requirements.txt
cv2
imutils
numpy
pytesseract
RPi.GPIO